* origin/pr/22917: (24 commits)
  Move (defaulted) implementation of Character into the cpp file.
  Fix pimpl class to better detect usage of copy constructor
  Remove unused include of "iosfwd" from "skill.h"
  Use Character::_skills directly in (de)serialization
  Remove unused function Skill::skill_count
  Use the size of the sorted skills list instead of count of all skills
  Iterate over existing skill levels instead of all skills.
  Iterate over existing skill levels instead of all skills.
  Iterate over existing skill levels instead of all skills.
  Iterate over existing skill levels instead of all skills.
  Extract commonly used expression into local reference.
  Replace usage of Skill::get with string_id<Skill>::obj
  Move game::rustCheck into the player class
  Remove SkillLevel::operator int()
  Remove unnecessary includes of "skill.h"
  Correct debug messages for invalid skills in "character.cpp":
  Change Character::get_skill_level to return an integer
  Add alternative functions to get SkillLevel object from Character:
  Rename Character::boost_skill_level to mod_skill_level
  Change Character::_skills to be a pimpl.
  ...

+++ src/crafting_gui.cpp
@@ -10,6 +10,7 @@
 #include "game.h"
 #include "translations.h"
 #include "string_formatter.h"
+#include "skill.h"
 #include "catacharset.h"
 #include "output.h"
 #include "json.h"
@@ -470,8 +471,7 @@ const recipe *select_crafting_recipe( int &batch_size )
                     mvwprintz( w_data, ypos++, xpos, col, _( "Your skill level: N/A" ) );
                 } else {
                     mvwprintz( w_data, ypos++, xpos, col, _( "Your skill level: %d" ),
-                               // Macs don't seem to like passing this as a class, so force it to int
-                               ( int )g->u.get_skill_level( current[line]->skill_used ) );
+                               g->u.get_skill_level( current[line]->skill_used ) );
                 }
 
                 const int turns = g->u.time_to_craft( *current[line], count ) / MOVES( 1 );
